requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML Libraries
openai==1.3.0
transformers==4.35.0
torch==2.1.0
huggingface-hub==0.19.0
sentence-transformers==2.2.2
vaderSentiment==3.3.2
textblob==0.17.1

# Web Interface
gradio==4.7.1
streamlit==1.28.0

# Dashboard & Analytics
plotly==5.17.0
dash==2.14.0
panel==1.3.0

# Database & Storage
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.25.0

# Communication APIs
twilio==8.10.0
requests==2.31.0

# Audio Processing (for Phase 2)
speechrecognition==3.10.0
pydub==0.25.1
gtts==2.4.0

# Deployment & Hosting
anvil-uplink==0.4.2
gunicorn==21.2.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
